requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
groq
httpx
